Addressed review feedback
- Kept only the shadow memory sample - the other one is only useful for heap that embedded users try to avoid.
- Moved the shadow memory to the end of RAM to avoid changing user code/layout too much.
- Replaced printf() with simple retargetable hooks.
- Added a bit of explanation what KASan is in the README.
