+++ arm-software/embedded/packagetest/cpp-baremetal-semihosting-kasan.test
@@ -1,0 +1,7 @@
+# RUN: make -C %samples_dir/src/cpp-baremetal-semihosting-kasan clean
+# RUN: not make -C %samples_dir/src/cpp-baremetal-semihosting-kasan run BIN_PATH=%unpack_directory/bin KASAN_TEST=1 > %t.out 2>&1
+# RUN: FileCheck %s < %t.out
+# RUN: make -C %samples_dir/src/cpp-baremetal-semihosting-kasan clean
+# CHECK: C++ KASan shadow-memory sample
+# CHECK: Test 1: automatic stack redzone
+# CHECK: KASAN: invalid store
